+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ New Features
 ============
 * `Update styling of switch component <https://github.com/anvilistas/anvil-extras/pull/56>`_
 * `Include pagination_click event in augment module <https://github.com/anvilistas/anvil-extras/pull/55>`_
+* `Slider Component <https://github.com/anvilistas/anvil-extras/pull/60>`_
 
 Changes
 =======